Create abstract
Railway is the most popular and friendly transportation system of the largest part of the cities in the world. Train is widely used for comfortable and safe journey in a reasonable fare. The train is facing unexpected situation in travelling because of wrong signal, wrong track switching, insecure level crossing etc. for which collision have been occurred. The major drawback of the accident takes there is no proper communication. In order to provide safe drive, wireless communication is used between the trains. We proposed a new system to communicate the trains through zigbee module. And also continuously monitor the train derailed or not. 
If any accident occurs, then the alert will be updated in IOT webpage with location immediately. This new innovative technology will increase the reliability of safety systems of railway. By implementing these features in real time application, we can save the human lives.